feat: 일급컬렉션 Sections 추가
