+++ src/main/java/nextstep/subway/advice/exception/LineNotFoundException.java
@@ -7,7 +7,6 @@ public class LineNotFoundException extends RuntimeException {
     }
 
     public LineNotFoundException(Long id) {
-        super("존재하는 노선이 없습니다 (id : " +id+" )");
+        super(String.format("존재하는 노선이 없습니다 (id : %d)", id));
     }
-
 }
