+++ src/main/java/nextstep/subway/advice/exception/StationNotFoundException.java
@@ -7,6 +7,6 @@ public class StationNotFoundException extends RuntimeException {
     }
 
     public StationNotFoundException(Long id) {
-        super("존재하는 역이 없습니다 (id : " +id+" )");
+        super(String.format("존재하는 역이 없습니다 (id : %d", id));
     }
 }
